Manage Lab Sessions
The project consists in developping an application in order to manage
lab sessions. The
project idea actually comes from a real need of professors in order to
be able to give marks to
students using simply their smartphones or tablets. This application
contains a mobile application (Portable on diffrent OSs) and also a web
one in order to give the user the choice to use it from a computer. Used
technologies : AngularJS, Bootstrap, Apache Cordova , Parse